requirements.txt
httpx==0.25.2
tenacity==8.2.2
beautifulsoup4==4.12.2
feedparser==6.0.10
python-dateutil==2.8.2
pydantic==1.10.12
structlog==23.1.0
ujson==5.8.0
rapidfuzz==3.5.2
tldextract==3.4.4
pytest==7.4.2
